rsx: Minor refactoring RSXThread
- Part 1 of many
