+++ rpcs3/Emu/RSX/RSXFIFO.h
@@ -32,6 +32,22 @@ namespace rsx
 			EMIT_BARRIER = 2
 		};
 
+		enum class state : u8
+		{
+			running = 0,
+			empty = 1,    // PUT == GET
+			spinning = 2, // Puller continuously jumps to self addr (synchronization technique)
+			nop = 3,      // Puller is processing a NOP command
+			lock_wait = 4,// Puller is processing a lock acquire
+			paused = 5,   // Puller is paused externallly
+		};
+
+		enum class interrupt_hint : u8
+		{
+			conditional_render_eval = 1,
+			zcull_sync = 2
+		};
+
 		struct register_pair
 		{
 			u32 reg;
