+++ rpcs3/Emu/RSX/RSXThread.h
@@ -28,6 +28,11 @@
 #include "Emu/IdManager.h"
 #include "Emu/system_config.h"
 
+#include "Core/RSXDisplay.h"
+#include "Core/RSXFrameBuffer.h"
+#include "Core/RSXIOMap.hpp"
+#include "Core/RSXVertexTypes.h"
+
 extern atomic_t<bool> g_user_asked_for_frame_capture;
 extern atomic_t<bool> g_disable_frame_limit;
 extern rsx::frame_trace_data frame_debug;
@@ -39,84 +44,6 @@ namespace rsx
 	{
 		class display_manager;
 	}
-
-	struct rsx_iomap_table
-	{
-		static constexpr u32 c_lock_stride = 8192;
-
-		std::array<atomic_t<u32>, 4096> ea;
-		std::array<atomic_t<u32>, 4096> io;
-		std::array<shared_mutex, 0x1'0000'0000 / c_lock_stride> rs;
-
-		rsx_iomap_table() noexcept;
-
-		// Try to get the real address given a mapped address
-		// Returns -1 on failure
-		u32 get_addr(u32 offs) const noexcept
-		{
-			return this->ea[offs >> 20] | (offs & 0xFFFFF);
-		}
-
-		template <bool IsFullLock, uint Stride>
-		bool lock(u32 addr, u32 len, cpu_thread* self = nullptr) noexcept
-		{
-			if (len <= 1) return false;
-			const u32 end = addr + len - 1;
-
-			bool added_wait = false;
-
-			for (u32 block = addr / c_lock_stride; block <= (end / c_lock_stride); block += Stride)
-			{
-				auto& mutex_ = rs[block];
-
-				if (IsFullLock ? !mutex_.try_lock() : !mutex_.try_lock_shared()) [[ unlikely ]]
-				{
-					if (self)
-					{
-						added_wait |= !self->state.test_and_set(cpu_flag::wait);
-					}
-
-					if (!self || self->id_type() != 0x55u)
-					{
-						IsFullLock ? mutex_.lock() : mutex_.lock_shared();
-					}
-					else
-					{
-						while (IsFullLock ? !mutex_.try_lock() : !mutex_.try_lock_shared())
-						{
-							self->cpu_wait({});
-						}
-					}
-				}
-			}
-
-			if (added_wait)
-			{
-				self->check_state();
-			}
-
-			return true;
-		}
-
-		template <bool IsFullLock, uint Stride>
-		void unlock(u32 addr, u32 len) noexcept
-		{
-			ensure(len >= 1);
-			const u32 end = addr + len - 1;
-
-			for (u32 block = (addr / 8192); block <= (end / 8192); block += Stride)
-			{
-				if constexpr (IsFullLock)
-				{
-					rs[block].unlock();
-				}
-				else
-				{
-					rs[block].unlock_shared();
-				}
-			}
-		}
-	};
 
 	enum framebuffer_creation_context : u8
 	{
@@ -175,22 +102,6 @@ namespace rsx
 		all_interrupt_bits = memory_config_interrupt | backend_interrupt | display_interrupt | pipe_flush_interrupt
 	};
 
-	enum FIFO_state : u8
-	{
-		running = 0,
-		empty = 1,    // PUT == GET
-		spinning = 2, // Puller continuously jumps to self addr (synchronization technique)
-		nop = 3,      // Puller is processing a NOP command
-		lock_wait = 4,// Puller is processing a lock acquire
-		paused = 5,   // Puller is paused externallly
-	};
-
-	enum FIFO_hint : u8
-	{
-		hint_conditional_render_eval = 1,
-		hint_zcull_sync = 2
-	};
-
 	enum result_flags: u8
 	{
 		result_none = 0,
@@ -205,264 +116,6 @@ namespace rsx
 		u32 col = __builtin_COLUMN(),
 		const char* file = __builtin_FILE(),
 		const char* func = __builtin_FUNCTION());
-
-	struct tiled_region
-	{
-		u32 address;
-		u32 base;
-		GcmTileInfo *tile;
-		u8 *ptr;
-
-		void write(const void *src, u32 width, u32 height, u32 pitch);
-		void read(void *dst, u32 width, u32 height, u32 pitch);
-	};
-
-	struct vertex_array_buffer
-	{
-		rsx::vertex_base_type type;
-		u8 attribute_size;
-		u8 stride;
-		std::span<const std::byte> data;
-		u8 index;
-		bool is_be;
-	};
-
-	struct vertex_array_register
-	{
-		rsx::vertex_base_type type;
-		u8 attribute_size;
-		std::array<u32, 4> data;
-		u8 index;
-	};
-
-	struct empty_vertex_array
-	{
-		u8 index;
-	};
-
-	struct draw_array_command
-	{
-		u32 __dummy;
-	};
-
-	struct draw_indexed_array_command
-	{
-		std::span<const std::byte> raw_index_buffer;
-	};
-
-	struct draw_inlined_array
-	{
-		u32 __dummy;
-		u32 __dummy2;
-	};
-
-	struct interleaved_attribute_t
-	{
-		u8 index;
-		bool modulo;
-		u16 frequency;
-	};
-
-	struct interleaved_range_info
-	{
-		bool interleaved = false;
-		bool single_vertex = false;
-		u32  base_offset = 0;
-		u32  real_offset_address = 0;
-		u8   memory_location = 0;
-		u8   attribute_stride = 0;
-
-		rsx::simple_array<interleaved_attribute_t> locations;
-
-		// Check if we need to upload a full unoptimized range, i.e [0-max_index]
-		std::pair<u32, u32> calculate_required_range(u32 first, u32 count) const;
-	};
-
-	enum attribute_buffer_placement : u8
-	{
-		none = 0,
-		persistent = 1,
-		transient = 2
-	};
-
-	class vertex_input_layout
-	{
-		int m_num_used_blocks = 0;
-		std::array<interleaved_range_info, 16> m_blocks_data{};
-
-	public:
-		rsx::simple_array<interleaved_range_info*> interleaved_blocks{};  // Interleaved blocks to be uploaded as-is
-		std::vector<std::pair<u8, u32>> volatile_blocks{};                // Volatile data blocks (immediate draw vertex data for example)
-		rsx::simple_array<u8> referenced_registers{};                     // Volatile register data
-
-		std::array<attribute_buffer_placement, 16> attribute_placement = fill_array(attribute_buffer_placement::none);
-
-		vertex_input_layout() = default;
-
-		interleaved_range_info* alloc_interleaved_block()
-		{
-			auto result = &m_blocks_data[m_num_used_blocks++];
-			result->attribute_stride = 0;
-			result->base_offset = 0;
-			result->memory_location = 0;
-			result->real_offset_address = 0;
-			result->single_vertex = false;
-			result->locations.clear();
-			result->interleaved = true;
-			return result;
-		}
-
-		void clear()
-		{
-			m_num_used_blocks = 0;
-			interleaved_blocks.clear();
-			volatile_blocks.clear();
-			referenced_registers.clear();
-		}
-
-		bool validate() const
-		{
-			// Criteria: At least one array stream has to be defined to feed vertex positions
-			// This stream cannot be a const register as the vertices cannot create a zero-area primitive
-
-			if (!interleaved_blocks.empty() && interleaved_blocks[0]->attribute_stride != 0)
-				return true;
-
-			if (!volatile_blocks.empty())
-				return true;
-
-			for (u8 index = 0; index < limits::vertex_count; ++index)
-			{
-				switch (attribute_placement[index])
-				{
-				case attribute_buffer_placement::transient:
-				{
-					// Ignore register reference
-					if (std::find(referenced_registers.begin(), referenced_registers.end(), index) != referenced_registers.end())
-						continue;
-
-					// The source is inline array or immediate draw push buffer
-					return true;
-				}
-				case attribute_buffer_placement::persistent:
-				{
-					return true;
-				}
-				case attribute_buffer_placement::none:
-				{
-					continue;
-				}
-				default:
-				{
-					fmt::throw_exception("Unreachable");
-				}
-				}
-			}
-
-			return false;
-		}
-
-		u32 calculate_interleaved_memory_requirements(u32 first_vertex, u32 vertex_count) const
-		{
-			u32 mem = 0;
-			for (auto &block : interleaved_blocks)
-			{
-				const auto range = block->calculate_required_range(first_vertex, vertex_count);
-				mem += range.second * block->attribute_stride;
-			}
-
-			return mem;
-		}
-	};
-
-	struct framebuffer_layout
-	{
-		ENABLE_BITWISE_SERIALIZATION;
-
-		u16 width;
-		u16 height;
-		std::array<u32, 4> color_addresses;
-		std::array<u32, 4> color_pitch;
-		std::array<u32, 4> actual_color_pitch;
-		std::array<bool, 4> color_write_enabled;
-		u32 zeta_address;
-		u32 zeta_pitch;
-		u32 actual_zeta_pitch;
-		bool zeta_write_enabled;
-		rsx::surface_target target;
-		rsx::surface_color_format color_format;
-		rsx::surface_depth_format2 depth_format;
-		rsx::surface_antialiasing aa_mode;
-		rsx::surface_raster_type raster_type;
-		u32 aa_factors[2];
-		bool ignore_change;
-	};
-
-	struct frame_statistics_t
-	{
-		u32 draw_calls;
-		u32 submit_count;
-
-		s64 setup_time;
-		s64 vertex_upload_time;
-		s64 textures_upload_time;
-		s64 draw_exec_time;
-		s64 flip_time;
-	};
-
-	struct display_flip_info_t
-	{
-		std::deque<u32> buffer_queue;
-		u32 buffer;
-		bool skip_frame;
-		bool emu_flip;
-		bool in_progress;
-		frame_statistics_t stats;
-
-		inline void push(u32 _buffer)
-		{
-			buffer_queue.push_back(_buffer);
-		}
-
-		inline bool pop(u32 _buffer)
-		{
-			if (buffer_queue.empty())
-			{
-				return false;
-			}
-
-			do
-			{
-				const auto index = buffer_queue.front();
-				buffer_queue.pop_front();
-
-				if (index == _buffer)
-				{
-					buffer = _buffer;
-					return true;
-				}
-			}
-			while (!buffer_queue.empty());
-
-			// Need to observe this happening in the wild
-			rsx_log.error("Display queue was discarded while not empty!");
-			return false;
-		}
-	};
-
-	class vblank_thread
-	{
-		std::shared_ptr<named_thread<std::function<void()>>> m_thread;
-
-	public:
-		vblank_thread() = default;
-		vblank_thread(const vblank_thread&) = delete;
-
-		void set_thread(std::shared_ptr<named_thread<std::function<void()>>> thread);
-
-		vblank_thread& operator=(thread_state);
-		vblank_thread& operator=(const vblank_thread&) = delete;
-	};
 
 	struct backend_configuration
 	{
@@ -493,6 +146,7 @@ namespace rsx
 		u64 tsc;
 	};
 
+	// TODO: This class is a mess, this needs to be broken into smaller chunks, like I did for RSXFIFO and RSXZCULL (kd)
 	class thread : public cpu_thread
 	{
 		u64 timestamp_ctrl = 0;
@@ -586,7 +240,7 @@ namespace rsx
 			atomic_t<u64> idle_time{ 0 };  // Time spent idling in microseconds
 			u64 last_update_timestamp = 0; // Timestamp of last load update
 			u64 FIFO_idle_timestamp = 0;   // Timestamp of when FIFO queue becomes idle
-			FIFO_state state = FIFO_state::running;
+			FIFO::state state = FIFO::state::running;
 			u32 approximate_load = 0;
 			u32 sampled_frames = 0;
 		}
@@ -736,7 +390,7 @@ namespace rsx
 		/**
 		 * Execute a backend local task queue
 		 */
-		virtual void do_local_task(FIFO_state state);
+		virtual void do_local_task(FIFO::state state);
 
 		virtual void emit_geometry(u32) {}
 
@@ -778,7 +432,7 @@ namespace rsx
 		// sync
 		void sync();
 		flags32_t read_barrier(u32 memory_address, u32 memory_range, bool unconditional);
-		virtual void sync_hint(FIFO_hint hint, reports::sync_hint_payload_t payload);
+		virtual void sync_hint(FIFO::interrupt_hint hint, reports::sync_hint_payload_t payload);
 		virtual bool release_GCM_label(u32 /*address*/, u32 /*value*/) { return false; }
 
 		std::span<const std::byte> get_raw_index_array(const draw_clause& draw_indexed_clause) const;
@@ -899,126 +553,4 @@ namespace rsx
 	{
 		return g_fxo->try_get<rsx::thread>();
 	}
-
-	template<bool IsFullLock = false, uint Stride = 128>
-	class reservation_lock
-	{
-		u32 addr = 0;
-		u32 length = 0;
-
-		inline void lock_range(u32 addr, u32 length)
-		{
-			if (!get_current_renderer()->iomap_table.lock<IsFullLock, Stride>(addr, length, get_current_cpu_thread()))
-			{
-				length = 0;
-			}
-
-			this->addr = addr;
-			this->length = length;
-		}
-
-	public:
-		reservation_lock(u32 addr, u32 length)
-		{
-			if (g_cfg.core.rsx_accurate_res_access &&
-				addr < constants::local_mem_base)
-			{
-				lock_range(addr, length);
-			}
-		}
-
-		reservation_lock(u32 addr, u32 length, bool setting)
-		{
-			if (setting)
-			{
-				lock_range(addr, length);
-			}
-		}
-
-		// Multi-range lock. If ranges overlap, the combined range will be acquired.
-		// If ranges do not overlap, the first range that is in main memory will be acquired.
-		reservation_lock(u32 dst_addr, u32 dst_length, u32 src_addr, u32 src_length)
-		{
-			if (g_cfg.core.rsx_accurate_res_access)
-			{
-				const auto range1 = utils::address_range::start_length(dst_addr, dst_length);
-				const auto range2 = utils::address_range::start_length(src_addr, src_length);
-				utils::address_range target_range;
-
-				if (!range1.overlaps(range2)) [[likely]]
-				{
-					target_range = (dst_addr < constants::local_mem_base) ? range1 : range2;
-				}
-				else
-				{
-					// Very unlikely
-					target_range = range1.get_min_max(range2);
-				}
-
-				if (target_range.start < constants::local_mem_base)
-				{
-					lock_range(target_range.start, target_range.length());
-				}
-			}
-		}
-
-		// Very special utility for batched transfers (SPU related)
-		template <typename T = void>
-		void update_if_enabled(u32 addr, u32 _length, const std::add_pointer_t<T>& lock_release = std::add_pointer_t<void>{})
-		{
-			// This check is not perfect but it covers the important cases fast (this check is only an optimization - forcing true disables it)
-			if (length && (this->addr / rsx_iomap_table::c_lock_stride != addr / rsx_iomap_table::c_lock_stride || (addr % rsx_iomap_table::c_lock_stride + _length) > rsx_iomap_table::c_lock_stride) && _length > 1)
-			{
-				if constexpr (!std::is_void_v<T>)
-				{
-					// See SPUThread.cpp
-					lock_release->release(0);
-				}
-
-				unlock();
-				lock_range(addr, _length);
-			}
-		}
-
-		void unlock(bool destructor = false)
-		{
-			if (length)
-			{
-				get_current_renderer()->iomap_table.unlock<IsFullLock, Stride>(addr, length);
-
-				if (!destructor)
-				{
-					length = 0;
-				}
-			}
-		}
-
-		~reservation_lock()
-		{
-			unlock(true);
-		}
-	};
-
-	class eng_lock
-	{
-		rsx::thread* pthr;
-	public:
-		eng_lock(rsx::thread* target)
-			:pthr(target)
-		{
-			if (pthr->is_current_thread())
-			{
-				pthr = nullptr;
-			}
-			else
-			{
-				pthr->pause();
-			}
-		}
-
-		~eng_lock()
-		{
-			if (pthr) pthr->unpause();
-		}
-	};
 }
